+++ cli/src/logging.rs
@@ -43,7 +43,7 @@ impl<T: fmt::Display> fmt::Display for Padded<T> {
         write!(f, "{: <width$}", self.value, width = self.width)
     }
 }
-fn colored_level<'a>(style: &'a mut Style, level: Level) -> StyledValue<'a, &'static str> {
+fn colored_level(style: &mut Style, level: Level) -> StyledValue<'_, &'static str> {
     match level {
         Level::Trace => style.set_color(Color::Magenta).value("TRACE"),
         Level::Debug => style.set_color(Color::Blue).value("DEBUG"),